Bathithu

+++ btth01/admin/author.php
@@ -70,7 +70,7 @@
                                 <a href="edit_author.php?id=<?php echo $row['ma_tgia'] ?>"><i class="fa-solid fa-pen-to-square"></i></a>
                             </td>
                             <td>
-                                <a href="process_delete_author.php?id=<?php echo $row['ma_tgia'] ?>"><i class="fa-solid fa-trash"></i></a>
+                                <a href="process_delete_author.php?id=<?php echo $row['ma_tgia'] ?>" onclick="return confirm('Bạn có muốn xoá tác giả này không?')">  <i class="fa-solid fa-trash"></i></a>
                             </td>
                         </tr>
                         <?php
